fix(review): address pre-PR adversarial review findings
Fixes from the Phase-C review fan-out over the branch:

- H1 (demo break): dev-token default roles were "User", but the portal authorizes
  against mgmt_case_def/mgmt_record_type/mgmt_form/client_case/client_task/client_record
  -> the dev user saw nothing. DevTokenController now defaults to that role set.
- H2 (prod guard): DevTokenWarning now FAILS FAST (refuses to start) when
  wks.auth.mode=dev-token is combined with the 'prod' profile, unless
  wks.auth.devtoken.allow-in-prod=true. Verified: minimal,prod -> app refuses to start.
- bpm selector honesty: C7EngineClient re-keyed from wks.bpm.engine.camunda.version
  onto wks.bpm.engine, so wks.bpm.engine=none alone fully disables Camunda (matches
  Camunda7ClientScan + NoopBpmEngineClient).
- L3: BearerTokenHandlerInputResolver.getAllowedOrigin is now null/empty-safe
  (was an unconditional get(0) -> NPE for any token lacking allowed-origins).
- DevTokenScan: gated @ConditionalOnProperty(wks.auth.mode=dev-token) (was
  unconditional) + javadoc corrected to describe its real (package-scan) activation.
- application-minimal.yml: add wks.tenancy.claim-name=org; keep canonical
  database.type=jpa (the wks.datastore.type alias is best-effort, not used in-profile).

Test isolation fix (root cause of a 49-error full-suite failure the review surfaced):
the JPA integration tests declared a nested @SpringBootConfiguration with a
SecurityContextTenantHolder stub inside com.wks.caseengine.* — which CaseEngineScan
component-scans — leaking a duplicate bean into every full-context test. Moved both
ITs to com.wks.it (outside all production @ComponentScan). Full rest-api suite green.

Verified end-to-end: minimal core boots on H2 (0 Mongo attempts), seeds 16/8/2,
dev token carries the 6 portal roles, create-case persists (200), prod guard trips.
