+++ apps/java/libraries/api-security/src/main/java/com/wks/api/security/devtoken/DevTokenWarning.java
@@ -11,7 +11,10 @@
  */
 package com.wks.api.security.devtoken;
 
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.core.env.Environment;
+import org.springframework.core.env.Profiles;
 import org.springframework.stereotype.Component;
 
 import jakarta.annotation.PostConstruct;
@@ -20,8 +23,11 @@ import lombok.extern.slf4j.Slf4j;
 /**
  * Loud startup warning emitted when the embedded dev-token issuer is active
  * ({@code wks.auth.mode=dev-token}). Mirrors the security-disabled warning
- * pattern. This mode mints self-signed JWTs and must never be used in
- * production.
+ * pattern. This mode mints self-signed JWTs and must never be used in production.
+ *
+ * <p>As a guard it FAILS FAST (refuses to start) if dev-token mode is combined
+ * with the {@code prod} Spring profile, unless explicitly overridden with
+ * {@code wks.auth.devtoken.allow-in-prod=true} (testing only).
  *
  * @author wks
  */
@@ -30,8 +36,24 @@ import lombok.extern.slf4j.Slf4j;
 @ConditionalOnProperty(name = "wks.auth.mode", havingValue = "dev-token")
 public class DevTokenWarning {
 
+	private final Environment environment;
+
+	private final boolean allowInProd;
+
+	public DevTokenWarning(Environment environment,
+			@Value("${wks.auth.devtoken.allow-in-prod:false}") boolean allowInProd) {
+		this.environment = environment;
+		this.allowInProd = allowInProd;
+	}
+
 	@PostConstruct
-	public void warn() {
+	public void warnOrFailFast() {
+		if (!allowInProd && environment.acceptsProfiles(Profiles.of("prod"))) {
+			throw new IllegalStateException(
+					"wks.auth.mode=dev-token must NEVER run under the 'prod' profile - refusing to start. "
+							+ "Use a real identity provider (wks.auth.mode=keycloak), or set "
+							+ "wks.auth.devtoken.allow-in-prod=true only for controlled testing.");
+		}
 		log.warn("==================================================================================");
 		log.warn("  DEV-TOKEN AUTHENTICATION IS ACTIVE (wks.auth.mode=dev-token)");
 		log.warn("  An embedded issuer is minting self-signed RS256 JWTs at /dev-auth/**.");
